+++ axum-core/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Unreleased
 
+- **added:** Add `IntoResponseHeaders` trait ([#644])
 - **breaking:** Using `HeaderMap` as an extractor will no longer remove the headers and thus
   they'll still be accessible to other extractors, such as `axum::extract::Json`. Instead
   `HeaderMap` will clone the headers. You should prefer to use `TypedHeader` to extract only the
@@ -20,7 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - The `HeadersAlreadyExtracted` variant has been removed from these rejections:
         - `RequestAlreadyExtracted`
         - `RequestPartsAlreadyExtracted`
-    - `<HeaderMap as FromRequest<_>>::Error` has been changed to `std::convert::Infallible`.
+    - `<HeaderMap as FromRequest<_>>::Rejection` has been changed to `std::convert::Infallible`.
 - **breaking:** `axum::http::Extensions` is no longer an extractor (ie it
   doesn't implement `FromRequest`). The `axum::extract::Extension` extractor is
   _not_ impacted by this and works the same. This change makes it harder to
@@ -31,10 +32,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - `RequestParts::extensions` returns `&Extensions`.
     - `RequestParts::extensions_mut` returns `&mut Extensions`.
     - `RequestAlreadyExtracted` has been removed.
-    - `<Request as FromRequest>::Error` is now `BodyAlreadyExtracted`.
-    - `<http::request::Parts as FromRequest>::Error` is now `Infallible`.
+    - `<Request as FromRequest>::Rejection` is now `BodyAlreadyExtracted`.
+    - `<http::request::Parts as FromRequest>::Rejection` is now `Infallible`.
     - `ExtensionsAlreadyExtracted` has been removed.
 
+[#644]: https://github.com/tokio-rs/axum/pull/644
 [#698]: https://github.com/tokio-rs/axum/pull/698
 [#699]: https://github.com/tokio-rs/axum/pull/699